fix(taskCalendar): move formatTime function to util
