+++ src/app/pages/home/components/task-calendar/task-calendar.component.ts
@@ -7,6 +7,7 @@ import {
 } from '@angular/core'
 
 import { Task } from '../../../../shared/models/task'
+import { formatTime } from '../../../../shared/utilities/time'
 
 @Component({
   selector: 'task-calendar',
@@ -31,25 +32,17 @@ export class TaskCalendarComponent implements OnChanges {
   }
 
   getStartTime(task: Task) {
-    return TaskCalendarComponent.formatTime(new Date(task.timestamp))
+    return formatTime(new Date(task.timestamp))
   }
 
   setCurrentTime() {
     const now = new Date()
-    this.currentTime = TaskCalendarComponent.formatTime(now)
+    this.currentTime = formatTime(now)
 
     // NOTE: Compare current time with the start times of the tasks and
     // find out in between which tasks it should be shown in the interface
     const todaysTasks = this.tasks.get(new Date().setUTCHours(0, 0, 0, 0))
     this.timeIndex = todaysTasks.findIndex(t => t.timestamp >= now.getTime())
-  }
-
-  static formatTime(date) {
-    const hour = date.getHours()
-    const min = date.getMinutes()
-    const hourStr = hour < 10 ? '0' + String(hour) : String(hour)
-    const minStr = min < 10 ? '0' + String(min) : String(min)
-    return hourStr + ':' + minStr
   }
 
   hasExtraInfo(warningStr) {
